refactor: improve face detection by using inheritance
*   **Inheritance:** Use inheritance to create a modular code structure.
*   **Base Class:** `CVModel` contains common methods and attributes.
*   **Derived Class:** `FaceDetectionModel` inherits from `CVModel` and implements face detection logic.
*   **Benefits:** Improved readability, maintainability, and easier future extensions.
